add guest func files
