+++ guest_function.go
@@ -1,0 +1,1 @@
+package wasify
